feat: Implement multi-step notification alert creation process
- Added new screens for selecting coins, addresses, and configuring alerts in the notification alert creation flow.
- Updated router to include new screens: AddNotificationAlertCoins, AddNotificationAlertAddresses, and AddNotificationAlertConfig.
- Refactored AddNotificationAlert to navigate through the new steps instead of handling all logic in a single component.
- Introduced a banner notification for alert activation time, dismissible by the user.
- Updated styles for new components and added necessary utility functions for address and coin handling.
